Move the javascript to app/javascript
This conforms to the prescribed project structure used by Shakapacker: https://github.com/shakacode/shakapacker\#configuration-and-code
