+++ app/components/embed/file_component.html.erb
@@ -20,5 +20,4 @@
   <%= render 'embed/metadata_panel', viewer: viewer %>
   <%= render 'embed/embed_this/file', viewer: viewer %>
   <%= render 'embed/footer/file', viewer: viewer %>
-  <script src="<%= asset_url('tree.js') %>"></script>
 </div>
