+++ app/javascript/packs/file.js
@@ -1,9 +1,10 @@
-import { CssInjection } from '../../assets/javascripts/modules/css_injection';
-import { EmbedThis } from '../../assets/javascripts/modules/embed_this';
-import { PopupPanels } from '../../assets/javascripts/modules/popup_panels';
-import { DownloadAll } from '../../assets/javascripts/modules/download_all';
-import { FilePreview } from '../../assets/javascripts/modules/file_preview';
-import CommonViewerBehavior from '../../assets/javascripts/modules/common_viewer_behavior';
+import { CssInjection } from '../src/modules/css_injection';
+import { EmbedThis } from '../src/modules/embed_this';
+import { PopupPanels } from '../src/modules/popup_panels';
+import { DownloadAll } from '../src/modules/download_all';
+import { FilePreview } from '../src/modules/file_preview';
+import CommonViewerBehavior from '../src/modules/common_viewer_behavior';
+import '../src/tree'
 
 document.addEventListener('DOMContentLoaded', () => {
   CssInjection.injectFontIcons();
