+++ app/javascript/packs/geo.js
@@ -1,9 +1,9 @@
 
-import { CssInjection } from '../../assets/javascripts/modules/css_injection';
-import { EmbedThis } from '../../assets/javascripts/modules/embed_this';
-import { PopupPanels } from '../../assets/javascripts/modules/popup_panels';
-import GeoViewer from '../../assets/javascripts/modules/geo_viewer';
-import CommonViewerBehavior from '../../assets/javascripts/modules/common_viewer_behavior';
+import { CssInjection } from '../src/modules/css_injection';
+import { EmbedThis } from '../src/modules/embed_this';
+import { PopupPanels } from '../src/modules/popup_panels';
+import GeoViewer from '../src/modules/geo_viewer';
+import CommonViewerBehavior from '../src/modules/common_viewer_behavior';
 
 import 'leaflet';
 import '../../../vendor/assets/javascripts/Leaflet.Control.Custom';
