+++ app/javascript/packs/m3.js
@@ -1,6 +1,6 @@
-import { CssInjection } from '../../assets/javascripts/modules/css_injection.js';
-import CommonViewerBehavior from '../../assets/javascripts/modules/common_viewer_behavior.js';
-import M3Viewer from '../../assets/javascripts/modules/m3_viewer.js';
+import { CssInjection } from '../src/modules/css_injection.js';
+import CommonViewerBehavior from '../src/modules/common_viewer_behavior.js';
+import M3Viewer from '../src/modules/m3_viewer.js';
 
 CssInjection.appendToHead();
 CommonViewerBehavior.initializeViewer(M3Viewer.init);
