+++ app/javascript/packs/pdf_viewer.js
@@ -1,11 +1,11 @@
 'use strict';
 
-import { CssInjection } from '../../assets/javascripts/modules/css_injection.js';
-import { EmbedThis } from '../../assets/javascripts/modules/embed_this.js';
-import { PopupPanels } from '../../assets/javascripts/modules/popup_panels.js';
-import CommonViewerBehavior from '../../assets/javascripts/modules/common_viewer_behavior.js';
-import PDFViewer from '../../assets/javascripts/modules/pdf_viewer.js';
-import Fullscreen from '../../assets/javascripts/modules/fullscreen.js';
+import { CssInjection } from '../src/modules/css_injection.js';
+import { EmbedThis } from '../src/modules/embed_this.js';
+import { PopupPanels } from '../src/modules/popup_panels.js';
+import CommonViewerBehavior from '../src/modules/common_viewer_behavior.js';
+import PDFViewer from '../src/modules/pdf_viewer.js';
+import Fullscreen from '../src/modules/fullscreen.js';
 
 $(document).ready(function() {
   CssInjection.injectFontIcons();
