+++ app/javascript/packs/virtex_3d_viewer.js
@@ -1,11 +1,11 @@
 'use strict';
 
-import { CssInjection } from '../../assets/javascripts/modules/css_injection.js';
-import CommonViewerBehavior from '../../assets/javascripts/modules/common_viewer_behavior.js';
-import { EmbedThis } from '../../assets/javascripts/modules/embed_this.js';
-import { PopupPanels } from '../../assets/javascripts/modules/popup_panels.js';
-import Fullscreen from '../../assets/javascripts/modules/fullscreen.js';
-import Virtex3DViewer from '../../assets/javascripts/modules/virtex_3d_viewer.js';
+import { CssInjection } from '../src/modules/css_injection.js';
+import CommonViewerBehavior from '../src/modules/common_viewer_behavior.js';
+import { EmbedThis } from '../src/modules/embed_this.js';
+import { PopupPanels } from '../src/modules/popup_panels.js';
+import Fullscreen from '../src/modules/fullscreen.js';
+import Virtex3DViewer from '../src/modules/virtex_3d_viewer.js';
 
 $(document).ready(function() {
   CssInjection.injectFontIcons();
