+++ app/javascript/src/components/CdlCopyright.jsx
