+++ app/javascript/src/components/embedMode.jsx
