+++ app/javascript/src/modules/common_viewer_behavior.js
