+++ app/javascript/src/modules/css_injection.js
