+++ app/javascript/src/modules/download_all.js
