+++ app/javascript/src/modules/download_panel.js
