+++ app/javascript/src/modules/file_preview.js
