+++ app/javascript/src/modules/fullscreen.js
