+++ app/javascript/src/modules/geo_viewer.js
