+++ app/javascript/src/modules/media_viewer.js
