+++ app/javascript/src/modules/pdf_viewer.js
