+++ app/javascript/src/modules/popup_panels.js
