+++ app/javascript/src/modules/thumb_slider.js
