+++ app/javascript/src/modules/virtex_3d_viewer.js
