+++ app/javascript/src/plugins/analyticsPlugin.js
