+++ app/javascript/src/plugins/cdlAuthPlugin.js
