+++ app/javascript/src/plugins/embedModePlugin.js
