+++ app/javascript/src/plugins/miradorZoomBugPlugin.jsx
