+++ app/javascript/src/plugins/shareMenuPlugin.jsx
