+++ app/javascript/src/templates/geo_sidebar.hbs
