+++ app/javascript/src/templates/index_map_info.hbs
